pyproject.toml
[project]
name = "file-upload-insights-platform"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.29",
  "pydantic-settings>=2.2",
  "python-jose[cryptography]>=3.3",
  "boto3>=1.34",
  "pytest>=8.0",
  "pydantic[email]>=2.11.7",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
  "asyncio: mark test to be run with asyncio",
]
[dependency-groups]
dev = [
    "types-python-jose>=3.5.0.20250531",
    "httpx>=0.28.1",
    "pytest-asyncio>=1.0.0",
    "boto3-stubs[s3]>=1.38.36",
]

[tool.mypy]
python_version        = "3.11"
ignore_missing_imports = true          # ← silence “missing stub” notes
plugins               = ["pydantic.mypy"]
